requirements.txt
streamlit
pypdf
langchain
langchain-community
langchain-huggingface
langchain-groq
faiss-cpu
sentence-transformers
python-dotenv